phase-c-0/pyproject.toml
[project]
name = "gamemind-probe"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.2.0",
    "pydantic>=2.12.5",
    "requests>=2.33.1",
]
